Add 'resource_aws_amplify_test.go'.
